Add files via upload
I added the functionality of the survey. It's not aesthetically pleasing yet, but I have implemented the javascript and local storage functionality. It saves the prompts and answers from each stage of the survey and removes previously selected prompts. Also linked the last part of the survey to the user_checkbox page